# src/config/openapi/invoice-management-v1.yaml
openapi: 3.0.3
info:
  title: Invoice Management API
  version: 1.1.0
  description: 
    Spec-first per Users, Invoices, Tax Profiles, con pagination/filtri avanzati,
    errori Problem Details e sicurezza (JWT bearer + API Key).
...
servers:
  - url: /
paths:
...
  /invoices:
    get:
      tags: [Invoices]
      operationId: getInvoices
      summary: restituisce una lista  di fatture (paginata, filtrabile)
      security: [ { bearerAuth: [] } ]
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
        - name: customerId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/InvoiceStatus' }
        - name: dateFrom
          in: query
          schema: { type: string, format: date-time }
          description: Filtro per data emissione (>=)
        - name: dateTo
          in: query
          schema: { type: string, format: date-time }
          description: Filtro per data emissione (<=)
        - name: invoiceNumber
          in: query
          schema: { type: string }
          description: Ricerca per codice fattura (contains/exact a scelta server)
        - name: amountMin
          in: query
          schema: { type: string }
          description: Importo minimo (grandTotal) come string/decimal
        - name: amountMax
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Lista fatture
          headers:
            X-Total-Count:
              schema: { type: integer }
              description: Totale elementi senza pagination
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedInvoices' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "404": { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Invoices]
      operationId: createInvoice
      summary: permette la creaziomne di una nuova fattura (solo ADMIN)
      security: [ { bearerAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InvoiceCreateRequest' }
      responses:
        "201":
          description: Creata nuova fattura con successo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatedId' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "404": { $ref: '#/components/responses/NotFound' }
        "409": { $ref: '#/components/responses/Conflict' }
  /invoices/{id}:
    get:
      tags: [Invoices]
      operationId: getInvoiceById
      summary: recupera fattura per id
      security: [ { bearerAuth: [] } ]
      parameters: [ { $ref: '#/components/parameters/id'}]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvoiceResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Invoices]
      operationId: updateInvoiceById
      summary: modifica fattura per id (volendo anche elementi della fattura. Solo ADMIN)
      security: [ { bearerAuth: [] } ]
      parameters: [ { $ref: '#/components/parameters/id' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InvoiceUpdateRequest' }
      responses:
        "200":
          description: Aggiornata fattura con successo
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvoiceResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
        "400": { $ref: '#/components/responses/BadRequest' }
    delete:
      tags: [Invoices]
      operationId: deleteInvoiceById
      summary: elimina fattura per id (solo ADMIN)
      security: [ { bearerAuth: [] } ]
      parameters: [ { $ref: '#/components/parameters/id' } ]
      responses:
        "204": { description: Cancellata fattura con successo }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT per autenticazione utente.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Chiave API per integrazioni server-to-server.
  parameters:
    id:
      in: path
      name: id
      required: true
      description: Identificatore risorsa (cryptato)
      schema:
        type: string
    page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    pageSize:
      in: query
      name: pageSize
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
  responses:
    BadRequest:
      description: Richiesta non valida
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ErrorDetails' }
    Unauthorized:
      description: Non autenticato o token invalido
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ErrorDetails' }
...
    NotFound:
      description: Risorsa non trovata
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ErrorDetails' }
    Conflict:
      description: Conflitto (es. email duplicata)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ErrorDetails' }
  schemas:
    ErrorDetails:
      type: object
      properties:
        message: { type: string, nullable: true }
        status:  { type: integer, nullable: true }
        detail:  { type: string, nullable: true}

    # Auth
...
    InvoiceResponse:
      type: object
      required: [idInvoice, invoiceNumber, issueDate, currency, status, subtotal, taxTotal, grandTotal, createdAt, updatedAt]
      properties:
        idInvoice: { type: string }
        invoiceNumber: { type: string }
        issueDate: { type: string, format: date-time }
        dueDate: { type: string, format: date-time, nullable: true }
        currency: { type: string, example: "EUR" }
        issuerUserId: { type: string }
        billToCustomerTaxProfileId: { type: string, nullable: true }
        billToCustomerId: { type: string, nullable: true }
        shipToCustomerTaxProfileId: { type: string, nullable: true }
        shipToCustomerId: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/InvoiceStatus' }
        subtotal: { type: string, example: "100.00" }
        taxTotal: { type: string, example: "22.00" }
        grandTotal: { type: string, example: "122.00" }
        paymentMethod: { $ref: '#/components/schemas/PaymentMethod' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        items: { type: array,  nullable: true, items: { $ref: '#/components/schemas/InvoiceItem' } }

    
 # === Invoice Items ===
    InvoiceItem:
      type: object
      required: [codProduct, description, quantity, unitPrice, taxRate, lineTotal]
      properties:
        codProduct: { type: string }
        description: { type: string }
        quantity: { type: string }
        unitPrice: { type: string }
        taxRate: { type: string }
        lineTotal: { type: string }
    InvoiceCreateRequest:
      type: object
      required: [invoiceNumber, issueDate, currency, subtotal, taxTotal, grandTotal]
      properties:
        invoiceNumber: { type: string }
        issueDate: { type: string, format: date-time }
        dueDate: { type: string, format: date-time, nullable: true }
        currency: { type: string }
        issuerUserId: { type: string }
        issuerTaxProfileId: { type: string }
        billToCustomerTaxProfileId: { type: string, nullable: true }
        billToCustomerId: { type: string, nullable: true }
        shipToCustomerTaxProfileId: { type: string, nullable: true }
        shipToCustomerId: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/InvoiceStatus' }
        subtotal: { type: string }
        taxTotal: { type: string }
        grandTotal: { type: string }
        paymentMethod: { $ref: '#/components/schemas/PaymentMethod' }
        items: { type: array,  nullable: true, items: { $ref: '#/components/schemas/InvoiceItem' } }
    InvoiceUpdateRequest:
      type: object
      properties:
        issueDate: { type: string, format: date-time }
        dueDate: { type: string, format: date-time, nullable: true }
        currency: { type: string }
        billToCustomerTaxProfileId: { type: string, nullable: true }
        billToCustomerId: { type: string, nullable: true }
        shipToCustomerTaxProfileId: { type: string, nullable: true }
        shipToCustomerId: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/InvoiceStatus' }
        subtotal: { type: string }
        taxTotal: { type: string }
        grandTotal: { type: string }
        paymentMethod: { $ref: '#/components/schemas/PaymentMethod' }
        items: { type: array,  nullable: true, items: { $ref: '#/components/schemas/InvoiceItem' } }
    InvoiceStatus:
      type: string
      enum: [DRAFT, ISSUED, PAID, CANCELED]
    PaymentMethod:
      type: string
      enum: [CARD, BANK_TRANSFER, CASH, OTHER]

    # Tax Profiles
...
    PaginatedInvoices:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/InvoiceResponse' } }
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...
    CreatedId:
      type: object
      required: [id]
      properties:
        id:
          type: string